Update types, build and documentation (#193)
* Update feature flags and documentation; disable default features and clarify opt-in requirements

* Update documentation to clarify feature requirements for viewer and renderer

* Update docs

* Update documentation for installation and API changes; clarify method requirements and improve descriptions

* Fix examples

* Refactor type handling for camera and contact force; update documentation for API changes

* Remove redundant unreachable code annotations in build script for clearer logic flow

* Refactor MuJoCo download logic in build script; improve error handling and comments

* Fix macos errors

* Compile error with auto download on macos

* Small fixes

* Update build script to rerun if DOCS_RS environment variable changes; enhance migration guide with default feature changes

* Add links = "mujoco" to Cargo.toml for native link deduplication

Prevents multiple crates from linking MuJoCo into the same binary.
Cargo enforces that only one crate can declare links = "mujoco",
avoiding duplicate symbol errors at link time.


* fix: recreate MjrContext on model change in viewer and renderer

When the model changes at runtime (detected via signature comparison),
the MjvScene, camera, and perturbation state are properly recreated,
but the MjrContext was not. This left stale GPU resources (textures,
meshes, heightfields, skins) from the old model in GPU memory, causing
the new model to render with incorrect or missing visual assets.

MuJoCo's own C++ viewer (simulate.cc) explicitly calls
mjr_freeContext + mjr_makeContext when loading a new model.

This commit adds MjrContext recreation in both:
- MjViewer::update_scene() (GL context already current from render())
- MjRenderer::try_sync() (GL context made current before recreation)


* Changelog and rules

* Fix docstring

---------

+++ examples/contact_forces.rs
@@ -131,7 +131,7 @@ fn main() {
         if ncon > 0 {
             let force_strs: Vec<String> = (0..ncon)
                 .map(|j| {
-                    let force = data.contact_force(j as u32);
+                    let force = data.contact_force(j);
                     // The first 3 components are the force (normal + 2 tangential);
                     // the last 3 are the torque.  We report the force magnitude.
                     let norm = (force[0] * force[0]
